+++ react/components/Table/Toolbar.js
@@ -133,14 +133,24 @@ class Toolbar extends PureComponent {
 
     const forcedColor = 'c-on-base'
 
+    const inputSearchAlone =
+      inputSearch &&
+      !isDensityVisible &&
+      !isFieldsVisible &&
+      !isDownloadVisible &&
+      !isUploadVisible &&
+      !isExtraActionsVisible &&
+      !isNewLineVisible
+
     return (
       <div
         id="toolbar"
         className={`mb5 flex flex-row w-100 ${
           isSearchBarVisible ? 'justify-between' : 'justify-end'
-        }`}>
+        }`}
+      >
         {inputSearch && (
-          <div className="w-40">
+          <div className={inputSearchAlone ? 'w-100 w-40-ns' : 'w-40'}>
             <InputSearch
               disabled={loading}
               {...inputSearch}
@@ -154,12 +164,14 @@ class Toolbar extends PureComponent {
               id="toggleDensity"
               title={density.buttonLabel}
               ref={this.densityBtnRef}
-              className="relative mh1">
+              className="relative mh1"
+            >
               <ButtonWithIcon
                 icon={
                   <span
                     className="c-on-base mh2"
-                    style={ICON_OPTICAL_COMPENSATION}>
+                    style={ICON_OPTICAL_COMPENSATION}
+                  >
                     <IconDensity size={MEDIUM_ICON_SIZE} />
                   </span>
                 }
@@ -173,17 +185,19 @@ class Toolbar extends PureComponent {
                   className={`absolute ${
                     density.alignMenu === 'right' ? 'right-0' : 'left-0'
                   } z-999 ba b--muted-4 br2 mt2 mh2`}
-                  style={BOX_SHADOW_STYLE}>
+                  style={BOX_SHADOW_STYLE}
+                >
                   <div className="w-100 b2 br2 bg-base">
                     <div
                       style={{ height: 3 * FIELDS_BOX_ITEM_HEIGHT }}
-                      className="overflow-auto">
+                      className="overflow-auto"
+                    >
                       {DENSITY_OPTIONS.map((key, index) => {
                         const isKeySelected = selectedDensity === key
                         return (
                           <div
                             key={index}
-                            className={`flex justify-between ph6 pv3 ${
+                            className={`TRX flex justify-between ph6 pv3 ${
                               isKeySelected ? 'b--emphasis' : 'b--transparent'
                             } pointer hover-bg-muted-5 bl bw1`}
                             onClick={() => {
@@ -191,9 +205,11 @@ class Toolbar extends PureComponent {
                               this.handleToggleBox('isDensityBoxVisible')
                               density.handleCallback &&
                                 density.handleCallback(key)
-                            }}>
+                            }}
+                          >
                             <span
-                              className={`w-100 ${isKeySelected ? 'fw5' : ''}`}>
+                              className={`w-100 ${isKeySelected ? 'fw5' : ''}`}
+                            >
                               {density[`${key}OptionLabel`]}
                             </span>
                           </div>
@@ -210,12 +226,14 @@ class Toolbar extends PureComponent {
               id="toggleFieldsBtn"
               title={fields.label}
               ref={this.fieldsBtnRef}
-              className="relative mh1">
+              className="relative mh1"
+            >
               <ButtonWithIcon
                 icon={
                   <span
                     className="c-on-base mh2"
-                    style={ICON_OPTICAL_COMPENSATION}>
+                    style={ICON_OPTICAL_COMPENSATION}
+                  >
                     <IconColumns size={MEDIUM_ICON_SIZE} />
                   </span>
                 }
@@ -228,32 +246,37 @@ class Toolbar extends PureComponent {
                 <div
                   className={`absolute ${
                     fields.alignMenu === 'right' ? 'right-0' : 'left-0'
-                  } z-999 ba b--muted-4 br2 mt2 mh2`}>
+                  } z-999 ba b--muted-4 br2 mt2 mh2`}
+                >
                   <div
                     className="w-100 b2 br2 bg-base"
                     style={{
                       ...BOX_SHADOW_STYLE,
                       width: FIELDS_BOX_WIDTH,
-                    }}>
+                    }}
+                  >
                     <div className="flex inline-flex bb b--muted-4 w-100 pl6 pv4">
                       <Button
                         variation="secondary"
                         size="small"
-                        onClick={onShowAllColumns}>
+                        onClick={onShowAllColumns}
+                      >
                         {fields.showAllLabel}
                       </Button>
                       <div className="mh4">
                         <Button
                           variation="secondary"
                           size="small"
-                          onClick={onHideAllColumns}>
+                          onClick={onHideAllColumns}
+                        >
                           {fields.hideAllLabel}
                         </Button>
                       </div>
                     </div>
                     <div
                       style={{ height: this.calculateFieldsBoxHeight() }}
-                      className="overflow-auto">
+                      className="overflow-auto"
+                    >
                       {Object.keys(schema.properties)
                         .filter(
                           field =>
@@ -263,7 +286,8 @@ class Toolbar extends PureComponent {
                           <div
                             key={index}
                             className="flex justify-between ph6 pv3 pointer hover-bg-muted-5"
-                            onClick={() => onToggleColumn(field)}>
+                            onClick={() => onToggleColumn(field)}
+                          >
                             <span className="w-70 truncate">
                               {schema.properties[field].title || field}
                             </span>
@@ -284,7 +308,8 @@ class Toolbar extends PureComponent {
               <ButtonWithIcon
                 icon={
                   <span
-                    className={`${download.disabled ? '' : forcedColor} mh2`}>
+                    className={`${download.disabled ? '' : forcedColor} mh2`}
+                  >
                     <IconDownload size={MEDIUM_ICON_SIZE} />
                   </span>
                 }
@@ -292,7 +317,8 @@ class Toolbar extends PureComponent {
                 disabled={download.disabled}
                 isLoading={download.isLoading}
                 size="small"
-                onClick={download.handleCallback}>
+                onClick={download.handleCallback}
+              >
                 {download.label && (
                   <span className={`${download.disabled ? '' : forcedColor}`}>
                     {download.label}
@@ -307,7 +333,8 @@ class Toolbar extends PureComponent {
                 icon={
                   <span
                     className={`${upload.disabled ? '' : forcedColor} mh2`}
-                    style={ICON_OPTICAL_COMPENSATION}>
+                    style={ICON_OPTICAL_COMPENSATION}
+                  >
                     <IconUpload size={HEAVY_ICON_SIZE} />
                   </span>
                 }
@@ -315,7 +342,8 @@ class Toolbar extends PureComponent {
                 disabled={upload.disabled}
                 isLoading={upload.isLoading}
                 size="small"
-                onClick={upload.handleCallback}>
+                onClick={upload.handleCallback}
+              >
                 {upload.label && (
                   <span className={`${upload.disabled ? '' : forcedColor}`}>
                     {upload.label}
@@ -355,7 +383,8 @@ class Toolbar extends PureComponent {
                     key="new-line-button"
                     icon={<IconPlus solid size={LIGHT_ICON_SIZE} />}
                     onClick={newLine.handleCallback}
-                    {...newLineButtonProps}>
+                    {...newLineButtonProps}
+                  >
                     {newLine.label}
                   </ButtonWithIcon>,
                   <ActionMenu
@@ -370,7 +399,8 @@ class Toolbar extends PureComponent {
               <ButtonWithIcon
                 icon={<IconPlus solid size={LIGHT_ICON_SIZE} />}
                 onClick={newLine.handleCallback}
-                {...newLineButtonProps}>
+                {...newLineButtonProps}
+              >
                 {newLine.label}
               </ButtonWithIcon>
             ))}